services/producer/requirements.txt
kafka-python==2.3.0
Faker==40.1.2
prometheus-client==0.20.0
